doc: Update README quickstart to fix issue with LangGraph.
The quickstart earlier used the minimal example from LangChain's doc, and worked since earlier we returned `StructuredTool` from Toolbox SDK. Now with recent changes, we return a subclass of `BaseTool` which caused issues with the way this quickstart binds the tools. This PR addresses that by using the updated create_react_agent from LangGraph, and also pretty prints to better explain the output to the end-user.