feat: Track utilized auth services (#203)
* fix: Add the no parameter check back again.

We will remove this once we actually implement the `strict` flag and centralize this functionality by moving this check to the tool's constructor in a future PR.

* fix: Reverse the error conditions to avoid masking of the second error.

* feat: Track utilized auth services

This PR expands the `identify_required_authn_params` helper function. Previously, it only identified missing auth params based on requirements and available token getters.

Now, the helper also returns a list of the auth token getters that were actually used during the identification process.

> [!NOTE]
> This enhancement is a preparatory step for implementing `strict` flag validation in an upcoming PR, allowing us to determine if all provided authentication methods were necessary.

* chore: Delint

* chore: Unpack tuple instead of index it

This makes it more readable.