chore(toolbox-core): Consolidate auth header creation logic (#213)
* chore: Add unit test coverage.

* chore: Consolidate auth header creation logic

Post adding the feature of adding client-level auth headers (#178), we have the logic for creating an auth header, from the given auth token getter name, in 3 different places.

This PR unifies all of that logic into a single helper to improve maintenance, and make it easier to change the way we add suffix/prefix, and reduces WET code.

* chore: Delint

* chore: Delint