docs: Enhance DEVELOPER.md for improved contribution guidelines (#316)
* docs: Enhance DEVELOPER.md for improved contribution guidelines

This PR significantly improves the `DEVELOPER.md` documentation across all Python SDK packages: `toolbox-core`, `toolbox-langchain`, and `toolbox-llamaindex`. The aim is to provide a more comprehensive, clear, and consistent guide for developers contributing to the project.

These updates aim to streamline the onboarding experience for new contributors and ensure that existing contributors have access to clear, up-to-date guidelines for developing and maintaining the SDKs.

* docs: Fix broken docs link in dev.md

* docs: Add pypi package links

* docs: Mention cloud build trigger in CI and Github actions section